ZED Blog/requirements.txt
Flask~=3.0.2
pymongo~=4.6.3
python-dotenv~=1.0.1